requirements.txt
numpy
pandas
scikit-learn
pytorch 
torchvision
tqdm

